improve flow derotation

+++ modules/OpticFlow/opticflow_code.h
@@ -7,8 +7,8 @@ extern unsigned int imgWidth, imgHeight;
 extern unsigned int tcp_port;
 extern unsigned int adjust_factor;
 extern unsigned int verbose;
-extern int opt_trans_x;
-extern int opt_trans_y;
+extern float opt_trans_x;
+extern float opt_trans_y;
 extern int opt_angle_x_raw;
 extern int opt_angle_y_raw;
